+++ apps/db/models/Trainer.py
@@ -7,7 +7,7 @@ class Trainer(db.Model):
     Primer_Apellido = db.Column(db.String(30), nullable=False)
     Segundo_Apellido = db.Column(db.String(30), nullable=False)
     Fecha_Nacimiento = db.Column(db.Date, nullable=False)
-    Edad = db.Column(db.Integer(11), nullable=False)
+    Edad = db.Column(db.Integer, nullable=False)
     Correo = db.Column(db.String(150), nullable=False)
     Telefono = db.Column(db.String(8), nullable=False)
     Estado = db.Column(db.Boolean, nullable=False, default=1)
